Add Scaffolding to generate tests from spec files
