+++ src/tools/ApiChecker/ApiChecker.csproj
@@ -1,18 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-    <ReleaseVersion>0.0.0</ReleaseVersion>
   </PropertyGroup>
-
+  <Import Project="../../Common/PackageInfo.props" />
   <ItemGroup>
     <None Remove="api.json" />
   </ItemGroup>
   <ItemGroup>
-    <Content Include="**/*">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    <Content Include="api.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="CommandLineParser" Version="2.9.0-preview1" />
+    <PackageReference Include="System.CodeDom" Version="5.0.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\PlaywrightSharp\PlaywrightSharp.csproj" />
